Added new directory, modules/legacy, and first legacy module, get-parameters. I'm going to stick things there that are removed from inxi but which we might want to use the logic of later, for example, the --option type handling is, though totally convoluted and not robust, still interesting, and might be useful for something or other.
Sample failure:

inxi -v 6 - does not, as you'd expect, give you verbosity level default and color 6, but rather complains about 
unsupported verbosity levels. In other words, the end user has to know a priori which order parameters can work in, which 
is silly and not at all user friendly.

We'll move all this to simple -c, --v getops handled args to make this uniform and easy to predict and use.