Added to branches the original forked infobash, the buggy one, with wrong 3.01 version number, gpl 3, failure to detect debian version string correctly.
This is the base we will work off of if there is ever any question re what the original code was.

To the best of my knowledge, this is the code we started with.

Also, for anyone else interested, now that slh owns infobash, and as of 3.04, changed the license to gpl 2, this will 
show you what code (all, basically, except maybe 2 lines) in infobash you can use with gpl 3.