(no version change, branch) Fine tuned the glx handling to open it up for all cases, and to also fix the bug, which I found the cause of now.
In case people don't like the glx information I'll add a further switch, but I think it's fine to show it for all cards now, especially 
with radeon and radeonhd work. Also improved direct rendering output to actually show now if -x or -F is used.

This will be inxi 1.0.8 once it's tested and confirmed good.