Syntax test fix for a broadcom wifi card id in -n output, failed to read path in /sys
Fixing this required using find, so find will have to be added to the recommended apps, but
leaving it off for now until sure this method is the only way to do it.

The earlier used grep method to get the paths did not work, why, I don't know.

This may require redoing the usb path finder logic, we'll see.