Adding xiin language file.  This file will load the US English version of output text xiin uses to communicate with the user.  This file does not determine the output to the debugger, that is dependent solely on the system being processed.  This is a preliminary upload, at this time, xiin doesn not load or read this file.
Future versions will first check for a network connection to the internet (without which, xiin is more or less useless anyhow).  Then it will find the language file per setting, US English as default.  The setting might be a CLI option or maybe a config file, or maybe it will check the current system to fine its locale.