creating gnubsd branch, this will for now hold the actual dev version of inxi that supports gnu bsd systems like kfreebsd debian, arch linux bsd, and so on. inxi itself will be that version, while the branches/bsd will be for true bsds like freebsd/openbsd
This lets me modify certain things like sed args, -r vs -R and so on in the real bsd branch, while leaving the dev branch for gnu
bsd as is. 

I'll be updating the trunk/inxi branch now and then to keep all of them more or less in sync, the goal being one master inxi
will then be the source for the bsd branches in real time always, via the small script inxi2bsd, which does the conversions. That 
script will be added to svn as well once it's cleaned up.