New version, new tarball. This update fixes a recent bug report with ancient dmidecode versions, that do not properly support the -s option. Now -M uses only one method for dmidecode, manual construction of the Machine data from the raw dmidecode file. The file output is also parsed a bit to make it more consistently reliable for inxi purposes.
This update also includes all recent bsd branch updates, including the new #!/usr/bin/env bash on top which
lets inxi run in any environment without changes. Also for bsd, sets sed -i/sed -i '' global value, which 
means that now all the branches are the same, except the bsd branches will contain the most recent tests
and bsd handling. 

As each step is reached, I'll release a new inxi that should be stable, this is the first one however that
can be used as is, no changes, for bsd, debian kfreebsd, and linux systems.

Pleasen note that most bsd features are either incomplete or missing completely at this point, but it's a
start.

Some initial changes as well to help options to show more correct linux or bsd terms. These will be updated
as time permits, it is a long process.