New version, new tarball, new man page. Unless disabled by distribution maintainers, offers weather -w option. With -x, -xx-, -xxx, shows more information. Basic line is just weather and system time there. -x adds time zone, which is useful for servers, particurly web servers. -x also adds wind speed. -xx adds humidity and barometric pressure. -xxx adds a possible new line, if data is available, heat index, wind chill, and dew point. -xxx also adds a line for location (blocked by irc/-z) / weather observation time.
-z filter applies as usual to location data, removes it in irc by default. -Z overrides override.

The api this uses is probably going to be dropped at some point, so this is just going to work 
while it works, then it will need to be updated at some point, so don't get very attached to it.

Also adds option to, with -w: -! location=<location string>
This lets users send an alternate location using either <city,state> or <postal code>
or <latitude,longitude> (commas for city,state and latitude,longitude are not optional, and the order
must be as listed.

If There is a developer flag if distro maintainers do not want this enabled, simply set:
B_ALLOW_WEATHER='false'
before packaging and the weather feature will be disabled.