New tarball, version, man page. Improved remote weather, now it uses -W, and deprecated -! location=..
That was too hard to type and too hard to remember. Also do more dyanamic reordering of weather
output, depending on how much data is present, and how many x options are used.

Added error handling for generic deprecated options, and for options that do not have the correct
syntax for OPTARG, like with -W.

This should about do it for the weather option for now unless I missed something somewhere.