New version, new man page, new tarball. Modified slightly -tc and -tm output to fix a pet peeve of mine. Now, if -I, -b, -F, or anything that can trigger the memory: used/total in Information line is not used, -tm will always show the system used/total ram data on the first line of the Memory item of -t output.
Also, if -xtc (trigger ram data in cpu output) is used, and -I is not triggered, and -tm is
not triggered, will also show system used/total ram data on the cpu first line.

I'd found it odd that this data did not appear when -tcm or -tm or -xtc were used, so this is
now fixed. I used the -t option a fair amount to find memory/cpu use issues, and usually I
don't use the option with other options, so the lack of total system ram data was odd.